Rename project from Nephelai to Aurboda
- Update all package names: nephelai → aurboda
- Rename Android package from se.hokasgard.nephelaiapp to net.aurboda
- Update database prefix from nephelai_{user} to aurboda_{user}
- Rename theme NephelaiAppTheme → AurbodaAppTheme
- Update API token name NEPHELIAI_API_TOKEN → AURBODA_API_TOKEN
- Update README with Norse mythology description of Aurboða
- Update all documentation references

Note: Existing user databases need to be renamed manually:
  sudo -u postgres psql -c "ALTER DATABASE nephelai_fiddur RENAME TO aurboda_fiddur;"
