+++ apps/android/README.md
@@ -1,6 +1,6 @@
-Nephelai App
-============
+Aurboda App
+===========
 
-Companion Android app for [Nepehai](https://github.com/fiddur/nephelai) Quantified Self aggregation.
+Companion Android app for [Aurboda](https://github.com/fiddur/aurboda) Quantified Self aggregation.
 
 Simple personal project to get Android Health Connect data from my phone into my database.
