+++ apps/android/app/src/main/java/net/aurboda/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package se.hokasgard.nephelaiapp.ui.theme
+package net.aurboda.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -34,7 +34,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun NephelaiAppTheme(
+fun AurbodaAppTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
