+++ apps/android/app/src/main/java/net/aurboda/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package se.hokasgard.nephelaiapp.ui.theme
+package net.aurboda.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
