+++ apps/android/app/src/test/java/net/aurboda/HealthDataModelsTest.kt
@@ -1,4 +1,4 @@
-package se.hokasgard.nephelaiapp
+package net.aurboda
 
 import org.junit.Test
 import org.junit.Assert.*
