+++ apps/backend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nephelai-backend",
+  "name": "aurboda-backend",
   "version": "1.0.0",
   "main": "src/api.ts",
   "scripts": {
