+++ apps/web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nephelai-web",
+  "name": "aurboda-web",
   "type": "module",
   "license": "AGPL-3.0-or-later",
   "author": "Fredrik Liljegren <fredrik@liljegren.org>",
