+++ docs/mcp-server.md
@@ -1,6 +1,6 @@
 # MCP Server
 
-Nephelai includes an MCP (Model Context Protocol) server that enables AI assistants like Claude to query health metrics and add manual tracking data.
+Aurboda includes an MCP (Model Context Protocol) server that enables AI assistants like Claude to query health metrics and add manual tracking data.
 
 ## Overview
 
@@ -205,9 +205,9 @@ The MCP server maintains per-user sessions:
 
 ## Connecting with Claude Desktop
 
-To connect Claude Desktop to your Nephelai MCP server:
+To connect Claude Desktop to your Aurboda MCP server:
 
-1. Start the Nephelai backend server
+1. Start the Aurboda backend server
 2. Obtain an authentication token via the login endpoint
 3. Configure Claude Desktop with the MCP server URL and Bearer token
 
@@ -216,7 +216,7 @@ Example Claude Desktop configuration (`claude_desktop_config.json`):
 ```json
 {
   "mcpServers": {
-    "nephelai": {
+    "aurboda": {
       "type": "http",
       "url": "http://localhost:3000/mcp",
       "headers": {
